requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
sqlalchemy
alembic
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
python-dotenv
python-multipart